// src/queries/model/paginationResponseDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * View Nest API
 * API documentation for View Nest application
 * OpenAPI spec version: 1.0
 */

export interface PaginationResponseDto {
  /** Number of items per page */
  limit: number;
  /** Current page number */
  page: number;
  /** Total number of items */
  total: number;
  /** Total number of pages */
  totalPages: number;
}